requirements.txt
streamlit
requests
python-dotenv
python-docx
pypdf
